pulled out shared drop function into helper
