+++ src/api/forms/repositories/form-definition-repository.js
@@ -1,6 +1,8 @@
 import Boom from '@hapi/boom'
 import { ObjectId } from 'mongodb'
 
+import { dropById } from './helpers.js'
+
 import { createLogger } from '~/src/helpers/logging/logger.js'
 import { db, DEFINITION_COLLECTION_NAME } from '~/src/mongo.js'
 
@@ -134,21 +136,7 @@ export async function get(formId, state = 'draft') {
 export async function drop(formId, session) {
   logger.info(`Deleting form definition with ID ${formId}`)
 
-  const coll = /** @satisfies {Collection<FormDefinition>} */ (
-    db.collection(DEFINITION_COLLECTION_NAME)
-  )
-
-  const result = await coll.deleteOne(
-    { _id: new ObjectId(formId) },
-    { session }
-  )
-  const { deletedCount } = result
-
-  if (deletedCount !== 1) {
-    throw new Error(
-      `Failed to delete form definition. Expected deleted count of 1, received ${deletedCount}`
-    )
-  }
+  await dropById(session, DEFINITION_COLLECTION_NAME, formId)
 
   logger.info(`Deleted form definition with ID ${formId}`)
 }
