+++ src/api/forms/repositories/form-metadata-repository.js
@@ -1,7 +1,9 @@
 import Boom from '@hapi/boom'
 import { MongoServerError, ObjectId } from 'mongodb'
 
-import { FormAlreadyExistsError } from './errors.js'
+import { FormAlreadyExistsError } from '../errors.js'
+
+import { dropById } from './helpers.js'
 
 import { createLogger } from '~/src/helpers/logging/logger.js'
 import { db, METADATA_COLLECTION_NAME } from '~/src/mongo.js'
@@ -166,21 +168,7 @@ export async function update(formId, update, session) {
 export async function drop(formId, session) {
   logger.info(`Deleting form metadata with ID ${formId}`)
 
-  const coll = /** @satisfies {Collection<FormMetadataDocument>} */ (
-    db.collection(METADATA_COLLECTION_NAME)
-  )
-
-  const result = await coll.deleteOne(
-    { _id: new ObjectId(formId) },
-    { session }
-  )
-  const { deletedCount } = result
-
-  if (deletedCount !== 1) {
-    throw new Error(
-      `Failed to delete form metadata. Expected deleted count of 1, received ${deletedCount}`
-    )
-  }
+  await dropById(session, METADATA_COLLECTION_NAME, formId)
 
   logger.info(`Deleted form metadata with ID ${formId}`)
 }
